Final patchset for Release v0.75

+++ rtpengine/centos/install.sh
@@ -85,7 +85,8 @@ fi
 
 # compile and install rtpengine from RPM's
 function install {
-    local RTPENGINE_RPM_VER TMP
+    local RTPENGINE_RPM_VER TMP BUILD_KERN_VERSIONS
+    local REBOOT_REQUIRED=0
     local OS_ARCH=$(uname -m)
     local OS_KERNEL=$(uname -r)
     local RHEL_BASE_VER=$(rpm -E %{rhel})
@@ -138,6 +139,7 @@ function install {
 
     if (( ${DISTRO_VER} >= 8 )); then
         dnf install -y kernel-devel-${OS_KERNEL} kernel-headers-${OS_KERNEL} || {
+            REBOOT_REQUIRED=1
             printwarn 'could not install kernel headers for current kernel'
             echo 'upgrading kernel and installing new headers'
             printwarn 'you will need to reboot the machine for changes to take effect'
@@ -145,6 +147,7 @@ function install {
         }
     else
         yum install -y kernel-devel-${OS_KERNEL} kernel-headers-${OS_KERNEL} || {
+            REBOOT_REQUIRED=1
             printwarn 'could not install kernel headers for current kernel'
             echo 'upgrading kernel and installing new headers'
             printwarn 'you will need to reboot the machine for changes to take effect'
@@ -156,6 +159,8 @@ function install {
         printerr "Could not install kernel headers"
         exit 1
     fi
+
+    BUILD_KERN_VERSIONS=$(joinwith '' ',' '' $(rpm -q kernel-headers | sed 's/kernel-headers-//g'))
 
     # rtpengine >= mr11.3.1.1 requires curl >= 7.43.0
     if versionCompare "$(tr -d '[a-zA-Z]' <<<"$RTPENGINE_VER")" gteq "11.3.1.1"; then
@@ -182,12 +187,6 @@ function install {
         fi
     fi
 
-    # create rtpengine user and group
-    # sometimes locks aren't properly removed (this seems to happen often on VM's)
-    rm -f /etc/passwd.lock /etc/shadow.lock /etc/group.lock /etc/gshadow.lock &>/dev/null
-    userdel rtpengine &>/dev/null; groupdel rtpengine &>/dev/null
-    useradd --system --user-group --shell /bin/false --comment "RTPengine RTP Proxy" rtpengine
-
     # reuse repo if it exists and matches version we want to install
     if [[ -d ${SRC_DIR}/rtpengine ]]; then
         if [[ "$(getGitTagFromShallowRepo ${SRC_DIR}/rtpengine)" != "${RTPENGINE_VER}" ]]; then
@@ -198,14 +197,21 @@ function install {
         git clone --depth 1 -c advice.detachedHead=false -b ${RTPENGINE_VER} https://github.com/sipwise/rtpengine.git ${SRC_DIR}/rtpengine
     fi
 
+    # apply our patches
+    (
+        cd ${SRC_DIR}/rtpengine &&
+        patch -p1 -N <${DSIP_PROJECT_DIR}/rtpengine/el-${RTPENGINE_VER}.patch
+    )
+    if (( $? > 1 )); then
+        printerr 'Failed patching RTPEngine files prior to build'
+        return 1
+    fi
+
     RTPENGINE_RPM_VER=$(grep -oP 'Version:.+?\K[\w\.\~\+]+' ${SRC_DIR}/rtpengine/el/rtpengine.spec)
     RPM_BUILD_ROOT="${HOME}/rpmbuild"
     rm -rf ${RPM_BUILD_ROOT} 2>/dev/null
     mkdir -p ${RPM_BUILD_ROOT}/SOURCES &&
     (
-        # some packages had to be compiled from source and therefore the default rpm build will fail
-        # we remove these from the the rpm spec files so we can still reliably install the other deps
-        # this also allows us to keep the standard post/pre build configurations from the spec file
         cd ${SRC_DIR} &&
         tar -czf ${RPM_BUILD_ROOT}/SOURCES/ngcp-rtpengine-${RTPENGINE_RPM_VER}.tar.gz \
             --transform="s%^rtpengine%ngcp-rtpengine-$RTPENGINE_RPM_VER%g" rtpengine/ &&
@@ -213,7 +219,7 @@ function install {
         # fix for BUG: "exec_prefix: command not found"
         function exec_prefix() { echo -n '/usr'; } && export -f exec_prefix &&
         # build the RPM's
-        rpmbuild -ba ${SRC_DIR}/rtpengine/el/rtpengine.spec &&
+        rpmbuild -ba --define "kversion $BUILD_KERN_VERSIONS" ${SRC_DIR}/rtpengine/el/rtpengine.spec &&
         rm -f ~/.rpmmacros && unset -f exec_prefix &&
         systemctl mask ngcp-rtpengine-daemon.service
 
@@ -234,18 +240,14 @@ function install {
         exit 1
     fi
 
-    # make sure RTPEngine kernel module configured
-    # skip if the kernel headers were not installed
-    if rpm -qa | grep -q "kernel-headers-$(uname -r)"; then
-        if [[ -z "$(find /lib/modules/${OS_KERNEL}/ -name 'xt_RTPENGINE.ko*' 2>/dev/null)" ]]; then
-            printerr "Problem installing RTPEngine kernel module"
-            exit 1
-        fi
+    # warn user if kernel module not loaded yet
+    if (( $REBOOT_REQUIRED == 1 )); then
+        printwarn "A reboot is required to load the RTPEngine kernel module"
     fi
 
     # ensure config dirs exist
     mkdir -p /var/run/rtpengine ${SYSTEM_RTPENGINE_CONFIG_DIR}
-    chown -R rtpengine:rtpengine /var/run/rtpengine
+    chown -R rtpengine:rtpengine /run/rtpengine
 
     # rtpengine config file
     # ref example config: https://github.com/sipwise/rtpengine/blob/master/etc/rtpengine.sample.conf
@@ -286,13 +288,11 @@ function install {
     cp -f ${DSIP_PROJECT_DIR}/resources/logrotate/rtpengine /etc/logrotate.d/rtpengine
 
     # Setup tmp files
-    echo "d /var/run/rtpengine.pid  0755 rtpengine rtpengine - -" > /etc/tmpfiles.d/rtpengine.conf
+    echo "d /run/rtpengine/rtpengine.pid  0755 rtpengine rtpengine - -" > /etc/tmpfiles.d/rtpengine.conf
 
     # Reconfigure systemd service files
     rm -f /lib/systemd/system/rtpengine.service 2>/dev/null
     cp -f ${DSIP_PROJECT_DIR}/rtpengine/systemd/rtpengine-v2.service /lib/systemd/system/rtpengine.service
-    cp -f ${DSIP_PROJECT_DIR}/rtpengine/rtpengine-{start-pre,stop-post} /usr/sbin/
-    chmod +x /usr/sbin/rtpengine-{start-pre,stop-post} /usr/bin/rtpengine
 
     # Reload systemd configs
     systemctl daemon-reload
